+++ gems/aws-sdk-core/spec/aws/json/error_handler_spec.rb
@@ -6,13 +6,13 @@ module Aws
   module Json
     describe ErrorHandler do
 
-      let(:client) {
+      let(:client) do
         ApiHelper.sample_json::Client.new(
           region: 'us-west-2',
           retry_limit: 0,
           credentials: Aws::Credentials.new('akid', 'secret')
         )
-      }
+      end
 
       let(:error_resp) { <<-JSON.strip }
 {
@@ -51,11 +51,11 @@ module Aws
           client.batch_get_item(
             request_items: {}
           )
-        }.to raise_error {|e|
+        }.to raise_error do |e|
           expect(e.code).to eq('ProvisionedThroughputExceededException')
           expect(e.message).to eq('foo')
           expect(e.foo).to eq('bar')
-        }
+        end
       end
 
       it 'ignore invalid characters when matching' do
@@ -65,11 +65,11 @@ module Aws
           client.batch_get_item(
             request_items: {}
           )
-        }.to raise_error {|e|
+        }.to raise_error do |e|
           expect(e.code).to eq('ProvisionedThroughputExceededException')
           expect(e.message).to eq('foo')
           expect(e.foo).to eq('bar')
-        }
+        end
       end
 
       it 'extracts code and message for unmodeled errors' do
@@ -79,10 +79,10 @@ module Aws
           client.batch_get_item(
             request_items: {}
           )
-        }.to raise_error {|e|
+        }.to raise_error do |e|
           expect(e.code).to eq('UnModeledException')
           expect(e.message).to eq('foo')
-        }
+        end
       end
 
       it 'extracts code and message for modeled empty struct errors' do
@@ -92,12 +92,71 @@ module Aws
           client.batch_get_item(
             request_items: {}
           )
-        }.to raise_error {|e|
+        }.to raise_error do |e|
           expect(e.code).to eq('EmptyStructError')
           expect(e.message).to eq('foo')
-        }
+        end
       end
 
+      describe 'AwsQueryCompatible' do
+        AwsQueryCompatibleClient = ApiHelper.sample_service(
+          api: {
+            'metadata' => {
+              'apiVersion' => '2018-11-07',
+              'awsQueryCompatible' => {},
+              'protocol' => 'json',
+              'endpointPrefix' => 'svc',
+              'signatureVersion' => 'v4'
+            },
+            'operations' => {
+              'Foo' => {
+                'name' => 'Foo',
+                'http' => { 'method' => 'POST', 'requestUri' => '/' },
+                'input' => { 'shape' => 'FooInput'}
+              }
+            },
+            'shapes' => {
+              'FooInput' => {
+                'type' => 'structure',
+                'members' => {}
+              },
+              'String' => { 'type' => 'string' }
+            }
+          }
+        ).const_get(:Client)
+
+        let(:client_aws_query_compatible) do
+          AwsQueryCompatibleClient.new(
+            region: 'us-west-2',
+            retry_limit: 0,
+            credentials: Aws::Credentials.new('akid', 'secret')
+          )
+        end
+
+        it 'extracts code and message from x-amzn-query-error' do
+          stub_request(:post, "https://svc.us-west-2.amazonaws.com/").
+            to_return(:status => 400, headers: {
+              'x-amzn-query-error': 'AwsQueryError;Sender'
+            }, :body => error_resp)
+          expect {
+            client_aws_query_compatible.foo()
+          }.to raise_error do |e|
+            expect(e.code).to eq('AwsQueryError')
+            expect(e.message).to eq('foo')
+          end
+        end
+
+        it 'fallback to default if missing x-amzn-query-error' do
+          stub_request(:post, "https://svc.us-west-2.amazonaws.com/").
+            to_return(:status => 400, headers: {}, :body => error_resp)
+          expect {
+            client_aws_query_compatible.foo()
+          }.to raise_error do |e|
+            expect(e.code).to eq('ProvisionedThroughputExceededException')
+            expect(e.message).to eq('foo')
+          end
+        end
+      end
     end
   end
 end