+++ gems/aws-sdk-core/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Feature - Support AwsQueryCompatible trait to read error code from x-amzn-query-error header.
+
 3.160.0 (2022-10-13)
 ------------------